+++ src/hooks/useUnifiedAlertsEnabled.ts
@@ -4,7 +4,7 @@ import { FeatureName } from 'types';
 
 import { useFeatureFlag } from './useFeatureFlag';
 
-export default function useUnifiedAlertsEnabled() {
+export function useUnifiedAlertsEnabled() {
   const { isEnabled: isEnabledByFF } = useFeatureFlag(FeatureName.UnifiedAlerting);
 
   return isEnabledByFF || config.unifiedAlertingEnabled;